# clif-backend: Eliminate FunctionEnvironment construction in feed_event()