@syrusakbary 2019-08-01
I'm following up with some of the changes that were done in this commit to move `FunctionBuilder` out of the `feed_event`: https://github.com/wasmerio/wasmer/commit/d440776bc080cd31086ef4779af84b1b0e0081e2