@syrusakbary 2019-08-01
After talking offline we decided to merge this PR and work on refactoring the `clif` backend so we can also move `FunctionBuilder` construction out of the `feed_event`